backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
fastmcp>=2.13.0
google-cloud-aiplatform>=1.76.0
google-cloud-storage>=2.18.0
google-auth>=2.36.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
langchain>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
sse-starlette>=2.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
langgraph-prebuilt>=0.1.0

